Create tomtesting.py
Python file which takes in IMU data from biopoint sensor, skips the first packet because it contains None type values, calculates the Pitch Roll and Yaw based on the quaternion values, and sends the data out over UDP as a list of 8 values (one for each packet received). Note - the executable needs a direct path so change yours accordingly! Also change the IP address and port number if needed